requirements.txt
alembic==1.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
greenlet==2.0.1
idna==3.4
load-dotenv==0.1.0
Mako==1.2.4
MarkupSafe==2.1.1
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
requests==2.28.1
sql==2022.4.0
SQLAlchemy==1.4.45
sqlparse==0.4.3
urllib3==1.26.13
XlsxWriter==3.0.3
